+++ src/main/java/org/realityforge/replicant/client/gwt/ReplicantConfig.java
@@ -1,0 +1,17 @@
+package org.realityforge.replicant.client.gwt;
+
+import org.realityforge.gwt.propertysource.client.PropertySource;
+import org.realityforge.gwt.propertysource.client.annotations.Namespace;
+
+@Namespace("replicant")
+public interface ReplicantConfig
+  extends PropertySource
+{
+  boolean shouldValidateRepositoryOnLoad();
+
+  boolean repositoryDebugOutputEnabled();
+
+  boolean subscriptionsDebugOutputEnabled();
+
+  boolean requestDebugOutputEnabled();
+}