+++ src/test/java/org/realityforge/replicant/client/transport/TestDataLoadService.java
@@ -1,6 +1,5 @@
 package org.realityforge.replicant.client.transport;
 
-import com.google.web.bindery.event.shared.SimpleEventBus;
 import java.util.Arrays;
 import java.util.LinkedList;
 import javax.annotation.Nonnull;
@@ -32,8 +31,7 @@ final class TestDataLoadService
            mock( EntityChangeBroker.class ),
            mock( EntityRepository.class ),
            mock( CacheService.class ),
-           mock( EntitySubscriptionManager.class ),
-           new SimpleEventBus() );
+           mock( EntitySubscriptionManager.class ) );
     _validator = mock( EntityRepositoryValidator.class );
   }